latihan unit 1.cpp

+++ Unit 1/latihan unit 1.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
  int menu;
-    float jarak, waktu, kecepatan;
+    int jarak, waktu, kecepatan;
 int main() {
